Millon Clinical Multiaxial Inventory-III (MCMI-III)

The MCMI-III data were originally collected by Rossi et al. (2010) and contain responses to 44 items from the Dutch version of the MCMI-III. The data has also been used by de la Torre et al. (2018) and Van der Ark et al. (2019) to demonstrate the applicability of diagnostic classification models in psychological assessment. The data set was uploaded to the Item Response Warehouse (Domingue et al., 2025) and is reformatted here.

Usage

mcmi_data

mcmi_qmatrix

Format

mcmi_data is a tibble containing response data with 1,208 rows and 45 variables:

id: Respondent identifier.

item.1-item.44: Dichotomous item responses to the 44 MCMI-III items.

mcmi_qmatrix is a tibble that identifies which disorders are measured by each item. This assessment contains 44 items measuring 3 disorders. The mcmi_qmatrix correspondingly is made up of 44 rows and 4 variables.

item: Item identifier, corresponds to item response columns in fraction_data.

Attribute columns: 4 columns, one for each attribute. Each is a dichotomous indicator for whether or not the disorder is measured by each item. A value of 1 indicates the disorder is measured by the item and a value of 0 indicates the disorder is not measured by the item.

Details

The attributes correspond to the presence of:

anxiety: Anxiety disorder

somatoform: Somatoform disorder

thought_disorder: Thought disorder

major_depression: Major depression

References

de la Torre, J., Van der Ark, L. A., & Rossi, G. (2018). Analysis of clinical data from cognitive diagnosis modeling framework. Measurement and Evaluation in Counseling and Development, 51(4), 281-296. doi:10.1080/07481756.2017.1327286

Domingue, B., Braginsky, M., Caffrey-Maffei, L., Gilbert, J. B., Kanopka, K., Kapoor, R., Lee, H., Liu, Y., Nadela, S., Pan, G., Zhang, L., Zhang, S., & Frank, M. C. (2025). An introduction to the Item Response Warehouse (IRW): A resource for enhancing data usage in psychometrics. Behavior Research Methods, 57, Article 276. doi:10.3758/s13428-025-02796-y

Rossi, G., Elklit, A., & Simonsen, E. (2010). Empirical evidence for a four factor framework of personality disorder organization: Multigroup confirmatory factor analysis of the Millon Clinical Multiaxial Inventory-III personality disorder scales across Belgian and Danish data samples. Journal of Personality Disorders, 24(1), 128-150. doi:10.1521/pedi.2010.24.1.128

Van der Ark, L. A., Rossi, G., & Sijtsma, K. (2019). Nonparametric item response theory and Mokken scale analysis, with relations to latent class models and cognitive diagnostic models. In M. von Davier & Y.-S. Lee (Eds.), Handbook of diagnostic classification models (pp. 21-45). Springer. doi:10.1007/978-3-030-05584-4_2